refactor: move config to internal
