+++ app.go
@@ -1,7 +1,7 @@
 package application
 
 import (
-	"echo-demo-project/config"
+	"echo-demo-project/internal/config"
 	"echo-demo-project/server"
 	"echo-demo-project/server/routes"
 	"log"
